+++ indico/core/signals/affiliations.py
@@ -12,11 +12,18 @@ _signals = Namespace()
 
 
 affiliation_created = _signals.signal('affiliation-created', '''
-Called from RHAffiliationAPI when processing new affiliations. The sender is the
+Called from RHAffiliationAPI when processing new affiliations. The *sender* is the
 Affiliation being created.
 ''')
 
 affiliation_updated = _signals.signal('affiliation-updated', '''
-Called from RHAffiliationAPI when processing updates. The sender is the
+Called from RHAffiliationAPI when processing updates. The *sender* is the
 Affiliation being updated and the update dict is passed via ``payload``.
+''')
+
+get_affiliation_filters = _signals.signal('get-affiliation-filters', '''
+Called when searching and validating affiliations in an affiliations field. The
+*sender* is the object from which affiliations are filtered. The current context
+is passed via ``context``. Expected to return a list of SQLAlchemy filters to
+restrict results.
 ''')