+++ indico/modules/events/registration/controllers/__init__.py
@@ -17,6 +17,7 @@ from indico.modules.events.registration.util import (check_registration_email, g
                                                      make_registration_schema, modify_registration,
                                                      process_registration_picture)
 from indico.modules.files.controllers import UploadFileMixin
+from indico.modules.users.models.affiliations import Affiliation
 from indico.util.i18n import _
 from indico.util.marshmallow import LowercaseString, UUIDString, not_empty
 from indico.web.args import use_kwargs
@@ -42,6 +43,11 @@ class RegistrationFormMixin:
             .options(form_items_strategy)
             .one()
         )
+
+    @property
+    def regform_uses_predefined_affiliations(self):
+        return (any(field.input_type == 'affiliation' for field in self.regform.active_fields) and
+                Affiliation.query.filter_by(is_deleted=False).has_rows())
 
 
 class RegistrationEditMixin:
@@ -87,6 +93,7 @@ class RegistrationEditMixin:
         return self.view_class.render_template(self.template_file, self.event,
                                                regform=self.regform,
                                                form_data=form_data,
+                                               has_predefined_affiliations=self.regform_uses_predefined_affiliations,
                                                payment_conditions=payment_event_settings.get(self.event, 'conditions'),
                                                payment_enabled=self.event.has_feature('payment'),
                                                registration=self.registration,